ai_news_agent/pyproject.toml
[project]
name = "ai_news_agent"
version = "0.1.0"
description = "AI 简报智能体 - 自动抓取 Reddit 论坛的 AI 最新动态并生成每日简报"
readme = "README.md"
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[tools]>=0.95.0",
    "praw>=7.7.0",
    "schedule>=1.2.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "duckduckgo-search>=8.1.1",
]

[project.scripts]
ai-news = "ai_news_agent.main:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ai_news_agent"]
